Extract OO git API to separate file
Also:
* Add to git fixtures ability to init with arbitrary name
* Fix git test regressions
