+++ rhino/tools/git/api.html
@@ -28,7 +28,7 @@ END LICENSE
 			scope.destroy();
 		</script>
 		<h1>Documentation is moving to <a href="../../../local/doc/typedoc/modules/slime.jrunscript.git.html">TypeScript</a></h1>
-		<div>See raw <a href="module.d.ts">TypeScript</a> code.</div>
+		<div>See raw <a href="module.fifty.ts">TypeScript</a> code.</div>
 		<div>
 			<h1>Context</h1>
 			<ul>
